fixed minor dispay error

+++ app/src/main/java/ch/band/inf2019/uk335/activities/OverviewActivity.java
@@ -45,7 +45,7 @@ public abstract class OverviewActivity  extends AppCompatActivity {
         } else {
             textViewFrequency.setText(getString(R.string.monthly));
         }
-        double cost =viewModel.getCostMonthYear()/100;
+        double cost = (double)viewModel.getCostMonthYear()/100;
         textViewSum.setText(NumberFormat.getCurrencyInstance(new Locale("DE","CH")).format(cost));
     }
 }
